pyproject.toml
[project]
name = "medicai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "couchbase>=4.0.0",
    "google-adk>=0.1.0",
    "google-genai>=0.1.0",
    "pytest>=7.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]
